correct response shape, cleanup, 670ms roundtrip
