+++ src/services/AxiosInstances.ts
@@ -1,8 +1,15 @@
 import axios, { AxiosInstance, AxiosResponse } from "axios";
+import {
+  formatDate,
+  getVolumeArrayFromOhlcv,
+  reshapeObject,
+} from "../utils/reshape";
+const _ = require("lodash")
+
 const alphaKey = process.env.APIKEY;
 
 //Axios instance for querying alphavantage api
-const alphavantage: AxiosInstance = axios.create({
+export const alphavantage: AxiosInstance = axios.create({
   baseURL: "https://www.alphavantage.co/query?function=CRYPTO_INTRADAY",
   timeout: 10000,
 });
@@ -26,5 +33,19 @@ export const getOhclv = async () => {
       console.log(data.data);
       return data;
     });
+    return data 
   //TODO: CATCH
 };
+
+//TODO: refactor
+alphavantage.interceptors.response.use((response) => {
+  console.log("interceptor");
+  let target = response.data["Time Series Crypto (5min)"];
+  let formattedDate = formatDate(target);
+  let objectValues = reshapeObject(target);
+  let volumeArray = getVolumeArrayFromOhlcv(objectValues);
+  let formattedOhcl = _.zip(formattedDate, objectValues);
+  let formattedData = { formattedOhcl, volumeArray }
+  
+  return formattedData 
+});
